requirements.txt
beautifulsoup4==4.9.2
bleach==3.2.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
decouple==0.0.7
dj-database-url==0.5.0
Django==2.2
django-extensions==3.0.9
django-heroku==0.3.1
djangorestframework==3.12.2
djongo==1.3.4
dnspython==2.1.0
gunicorn==20.0.4
idna==2.10
install==1.3.3
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
packaging==20.4
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.2.4
urllib3==1.26.3
webencodings==0.5.1
whitenoise==5.1.0
